feat(test): add unit test
