+++ packages/code-du-travail-frontend/src/lib/url.ts
@@ -4,7 +4,7 @@ const {
   publicRuntimeConfig: { AZURE_BASE_URL, AZURE_CONTAINER },
 } = getConfig();
 
-export function toUrl(file) {
+export function toUrl(file?: any): any {
   try {
     const [filename] = file.match(/[^/]+$/);
     return `${AZURE_BASE_URL}/${AZURE_CONTAINER}/${filename}`;
@@ -12,3 +12,8 @@ export function toUrl(file) {
     return file;
   }
 }
+
+export const removeQueryParameters = (url: string): string => {
+  const index = url.indexOf("?");
+  return index !== -1 ? url.substring(0, index) : url;
+};
